feat: Uso LocalStorage e reestruturação
- Uso do localStorage ao invés de uma lista fixa
- Desmembrado o arquivo app.js em arquivos que façam mais sentido
